Combine api.py and interface.py
Now hardware/builder.py.
